Raise if a contract is configured with a non-existant field
Prior to this change, if a user typoed the name of a field in a contract
declaration (or a developer typoed it in a unit test...) then the
contract would just ignore the configuration field.

This change raises an explicit error if a contract is passed a
configuration option it does not expect.