Allow Click 8.x.x
Click 8.0.0 is released 2021-05-12 (and 8.0.1 later). It includes type annotations and other useful features. However, import-linter restricts the versions to `<8`, which breaks any environments that would like to use Click 8.x.x — despite it works quite fine with this new version.

This PR allows Click 8.x.x to be used and raises the version limit to 9.x.x (not yet existent).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>