Show multiple errors when popping imports
Prior to this change, if we tried to pop multiple imports from the graph
that did not exist, we would only raise an error for the first one. This
meant that if we have multiple incorrect ignores for a contract, we will
only see one of them each time we lint imports (which is a bad
experience if you are trying to add a big list of potential ignores and
whittle them down).

This change raises a single error containing all of the missing imports.
We sort the imports alphabetically to make things easier.