Switch from toml to tomli/tomllib
The toml package is unmaintained, so it was appropriate to move away
from it as a dependency. The tomli package is the most appropriate
alternative, because it was incorporated into the Python standard
library as of Python 3.11, under the name tomllib.

Due to the prevalence of tomli as a compulsory dependency of many other
Python code quality tools (such as black, mypy, pytest and pylint),
there's a big chance that users will already have it installed if
they're using Import Linter. Therefore, tomli is now a compulsory
dependency (if you're using versions of Python earlier than 3.11).

This change has several benefits:

- Reduces the overall number of dependencies in the average Python
  application
- Removes a dependency on an unmaintained project
- Reduces the complexity of configuration-related testing
- Greatly simplifies the tox configuration