Handle '-1' line numbers in contract rendering code
As part of an important improvement to type safety in other parts of the
codebase, we can no longer pretend that it's okay for line numbers to be
strings. Previously, this was being done to magically encode question
marks into error information displayed to users.

Now we encode this data as an actual number - a -1, something that can
never be a valid line number by itself. This means the code that
displays these line numbers to users needs to be adjusted, which is what
this code does. We check for these values in a dedicated formatting
function, and replace them with question marks as necessary.