Fix pre-commit hook
Import Linter needs to be able to import the packages under test, which
will usually be the project's package. This doesn't seem to be easy to
achieve using language: python.

We also remove the limit on types to Python, since otherwise it won't
pick up on changes to contracts.

This could cause issues for devs who commit through a Git UI, since the
env won't be activated unless specially configured.